rename: topic exception 패키지 이동
