+++ src/main/java/com/tierlist/tierlist/topic/application/domain/service/TopicFavoriteService.java
@@ -3,9 +3,9 @@ package com.tierlist.tierlist.topic.application.domain.service;
 import com.tierlist.tierlist.member.application.domain.exception.MemberNotFoundException;
 import com.tierlist.tierlist.member.application.domain.model.Member;
 import com.tierlist.tierlist.member.application.port.out.persistence.MemberRepository;
+import com.tierlist.tierlist.topic.application.domain.exception.TopicNotFoundException;
 import com.tierlist.tierlist.topic.application.domain.model.Topic;
 import com.tierlist.tierlist.topic.application.domain.model.TopicFavorite;
-import com.tierlist.tierlist.topic.application.exception.TopicNotFoundException;
 import com.tierlist.tierlist.topic.application.port.in.service.TopicFavoriteUseCase;
 import com.tierlist.tierlist.topic.application.port.out.persistence.TopicFavoriteRepository;
 import com.tierlist.tierlist.topic.application.port.out.persistence.TopicRepository;
