+++ src/test/java/com/tierlist/tierlist/tierlist/application/domain/service/TierlistCreateServiceTest.java
@@ -16,8 +16,8 @@ import com.tierlist.tierlist.support.topic.FakeTopicRepository;
 import com.tierlist.tierlist.tierlist.application.domain.model.Tierlist;
 import com.tierlist.tierlist.tierlist.application.domain.model.command.TierlistCreateCommand;
 import com.tierlist.tierlist.tierlist.application.port.out.persistence.TierlistRepository;
+import com.tierlist.tierlist.topic.application.domain.exception.TopicNotFoundException;
 import com.tierlist.tierlist.topic.application.domain.model.Topic;
-import com.tierlist.tierlist.topic.application.exception.TopicNotFoundException;
 import com.tierlist.tierlist.topic.application.port.out.persistence.TopicRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
