+++ src/test/java/com/tierlist/tierlist/topic/application/domain/service/TopicCreateServiceTest.java
@@ -8,9 +8,9 @@ import com.tierlist.tierlist.category.application.domain.model.Category;
 import com.tierlist.tierlist.category.application.port.out.persistence.CategoryRepository;
 import com.tierlist.tierlist.support.category.FakeCategoryRepository;
 import com.tierlist.tierlist.support.topic.FakeTopicRepository;
+import com.tierlist.tierlist.topic.application.domain.exception.TopicDuplicationException;
 import com.tierlist.tierlist.topic.application.domain.model.Topic;
 import com.tierlist.tierlist.topic.application.domain.model.command.TopicCreateCommand;
-import com.tierlist.tierlist.topic.application.exception.TopicDuplicationException;
 import com.tierlist.tierlist.topic.application.port.out.persistence.TopicRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
