+++ src/test/java/com/tierlist/tierlist/topic/application/domain/service/TopicFavoriteServiceTest.java
@@ -12,9 +12,9 @@ import com.tierlist.tierlist.support.category.FakeCategoryRepository;
 import com.tierlist.tierlist.support.member.FakeMemberRepository;
 import com.tierlist.tierlist.support.topic.FakeTopicFavoriteRepository;
 import com.tierlist.tierlist.support.topic.FakeTopicRepository;
+import com.tierlist.tierlist.topic.application.domain.exception.TopicNotFoundException;
 import com.tierlist.tierlist.topic.application.domain.model.Topic;
 import com.tierlist.tierlist.topic.application.domain.model.TopicFavorite;
-import com.tierlist.tierlist.topic.application.exception.TopicNotFoundException;
 import com.tierlist.tierlist.topic.application.port.out.persistence.TopicFavoriteRepository;
 import com.tierlist.tierlist.topic.application.port.out.persistence.TopicRepository;
 import java.util.Optional;
